De-entangle code
