+++ packages/alfa-selector/src/selector/compound.ts
@@ -96,80 +96,6 @@ export namespace Compound {
   }
 
   /**
-   * Check if an element matches a compound selector containing a`::slotted` pseudo-element.
-   *
-   * @remarks
-   * `::slotted` can be used at the end of a compound selector, in which case the
-   * start of the compound selector must match the assigned slot.
-   * E.g., `div.foo::slotted(.bar)` matches a `.bar` slotted in a `div.foo`.
-   *
-   * @privateRemarks
-   * This must be defined here to avoid circular dependencies:
-   * Slotted -> Compound -> Simple -> Slotted.
-   * Yet, this may not fully live at the top-level Selector due to
-   * Complex.matches requiring to call this. For the sake of simplicity,
-   * we re-export that from Selector.
-   *
-   * @internal
-   */
-  export function matchSlotted(
-    element: Element & Slotable,
-    selector: Compound | Simple,
-    context: Context = Context.empty(),
-  ): boolean {
-    // The part of `selector` that must match the slot.
-    let slotSelector: Option<Compound> = None;
-    // The part of `selector` that must match the slotted element
-    // (pseudo-classes after ::slotted).
-    let qualifier: Option<Compound> = None;
-
-    const selectors = Compound.isCompound(selector)
-      ? [...selector.selectors]
-      : [selector];
-
-    if (selectors.filter(Slotted.isSlotted).length !== 1) {
-      // There is either 0, or several `::slotted()` in the compound selector.
-      return false;
-    }
-
-    // We know there is exactly one `::slotted()` in the compound selector.
-    const actualSelector = selectors.find(Slotted.isSlotted) as Slotted;
-
-    if (selectors.length > 1) {
-      // The slot selector is the start of the compound selector, until ::slotted.
-      slotSelector = Option.of(
-        Compound.of(
-          ...Iterable.takeUntil(selectors, (selector) =>
-            Slotted.isSlotted(selector),
-          ),
-        ),
-      );
-
-      qualifier = Option.of(
-        Compound.of(
-          ...Iterable.takeLastUntil(selectors, (selector) =>
-            Slotted.isSlotted(selector),
-          ),
-        ),
-      );
-    }
-
-    const slot = element.assignedSlot();
-
-    return (
-      // `element` must be slotted.
-      slot.some((slot) =>
-        // The slot must match the slot selector, if any.
-        slotSelector.every((selector) => selector.matches(slot, context)),
-      ) &&
-      // `element` must match the argument of the actual ::slotted selector.
-      actualSelector.selector.matches(element, context) &&
-      // `element` must match the qualifier, if any.
-      qualifier.every((selector) => selector.matches(element, context))
-    );
-  }
-
-  /**
    * {@link https://drafts.csswg.org/selectors/#typedef-compound-selector}
    *
    * @internal
